+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- **CUMULUS-4189**
+  - Added search button and removed  dispatching search/url update on every keystroke
 - **CUMULUS-4451**
   - Disable sort executions by name to avoid slow query in Postgres
   - Set default sort order for executions to -updatedAt to avoid slow query in Postgres